requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.1
Django==3.2.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
marshmallow==3.13.0
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
